simulator/backend/requirements.txt
requests==2.32.3
cryptography==42.0.5
bsdiff4==1.2.4
paho-mqtt
grpcio
grpcio-tools
protobuf
